Move EnsimeClient to a new client.py file
ensime.py is big enough to be unwieldy, and as I break apart
EnsimeClient to be less monolithic, the separation helps me think about
the distinctions of responsibilities for EnsimeClient versus Ensime.
