--- ensime_shared/spec/.ensime_cache/http
@@ -1,1 +1,0 @@
-42
